Address Samir's review
